+++ crates/credential/Cargo.toml
@@ -35,6 +35,12 @@ nebula-metadata = { path = "../metadata" }
 # Schema definitions
 nebula-schema = { path = "../schema" }
 
+# Storage port — RefreshClaimStore trait (ADR-0092 step 4A.3)
+nebula-storage-port = { path = "../storage-port" }
+
+# Metrics — pre-bound handles for refresh coordinator series
+nebula-metrics = { path = "../metrics" }
+
 # Resilience (circuit breaker, retry)
 nebula-resilience = { path = "../resilience" }
 
@@ -61,6 +67,16 @@ semver = { workspace = true }
 
 # Synchronization
 arc-swap = { workspace = true }
+parking_lot = { workspace = true }
+
+# LRU cache for per-credential circuit breakers in L1RefreshCoalescer
+lru = { workspace = true }
+
+# scopeguard — RAII cleanup for L2 claim + heartbeat teardown
+scopeguard = "1"
+
+# futures — select!/stream combinators in refresh coordinator
+futures = { workspace = true }
 
 # Registry primitives (Tech Spec §3.1 — KEY-keyed CredentialRegistry)
 # `ahash` 0.8 — ~3× faster than default SipHash at credential-key sizes
