+++ crates/credential/src/runtime/mod.rs
@@ -8,11 +8,17 @@
 pub mod dispatchers;
 pub mod executor;
 pub mod lease;
+pub mod refresh;
 pub mod scoped_accessor;
 
 pub use dispatchers::{dispatch_release, dispatch_revoke, dispatch_test};
 pub use executor::{ExecutorError, ResolveResponse, execute_continue, execute_resolve};
 pub use lease::{
     LeaseLifecycle, LeaseLifecycleConfig, LeaseLifecycleError, LeaseToken, RenewalPolicy,
+};
+pub use refresh::{
+    ConfigError, ReclaimSweepHandle, RefreshAttempt, RefreshConfigError, RefreshCoordConfig,
+    RefreshCoordMetrics, RefreshCoordinator, RefreshError, SentinelDecision,
+    SentinelThresholdConfig, SentinelTrigger,
 };
 pub use scoped_accessor::ScopedCredentialAccessor;
