+++ crates/credential/src/runtime/refresh/coordinator.rs
@@ -1,0 +1,1015 @@
+//! Outer two-tier refresh coordinator.
+//!
+//! See `docs/INTEGRATION_MODEL.md` for the two-tier refresh diagram, parameter invariants, and
+//! contention backoff.
+//!
+//! `RefreshCoordinator` composes:
+//!
+//! - **L1** -- `super::l1::L1RefreshCoalescer` (in-process oneshot coalesce
+//!   + per-credential circuit breaker + global concurrency semaphore).
+//! - **L2** -- `Arc<dyn nebula_storage_port::store::RefreshClaimStore>` (durable CAS-based claim
+//!   with TTL + heartbeat).
+//!
+//! Callers invoke `refresh_coalesced(credential_id, do_refresh)`. The
+//! coordinator acquires L1 first (fast in-process coalesce), then a
+//! durable L2 claim with contention backoff, runs the user's refresh
+//! closure under both locks, and releases both on the way out.
+
+use std::{
+    fmt,
+    future::Future,
+    sync::Arc,
+    time::{Duration, Instant},
+};
+
+use nebula_core::CredentialId;
+use nebula_storage_port::store::{
+    ClaimAttempt, ClaimToken, HeartbeatError, RefreshClaim, RefreshClaimError as RepoError,
+    RefreshClaimStore as RefreshClaimRepo, ReplicaId,
+};
+use tokio_util::sync::CancellationToken;
+use tracing::Instrument;
+
+use crate::audit::AuditSink;
+
+use super::{
+    audit::emit_claim_acquired,
+    l1::{L1RefreshCoalescer, RefreshAttempt as L1Attempt},
+    metrics::RefreshCoordMetrics,
+};
+
+// ──────────────────────────────────────────────────────────────────────────
+// Configuration
+// ──────────────────────────────────────────────────────────────────────────
+
+/// Configuration knobs for the two-tier coordinator.
+///
+/// Per sub-spec the four time-related parameters carry interlocking
+/// invariants verified by [`RefreshCoordConfig::validate`]:
+///
+/// - `heartbeat_interval × 3 <= claim_ttl` -- three heartbeat ticks must fit inside one claim TTL
+///   so two consecutive missed heartbeats still leave the claim valid until the next tick.
+/// - `refresh_timeout + 2 × heartbeat_interval <= claim_ttl` -- the holder must finish (or time
+///   out) before its claim can expire.
+/// - `reclaim_sweep_interval <= claim_ttl` -- sweeps must run at least as often as a claim's TTL
+///   so a crashed holder is reclaimed within one TTL window.
+///
+/// The boundary case `heartbeat_interval × 3 == claim_ttl` is allowed
+/// (mirrors the execution-lease shape: `ttl / 3 ==
+/// heartbeat_interval`).
+///
+/// CI test asserts `RefreshCoordConfig::default().validate().is_ok()`.
+#[derive(Clone, Debug)]
+pub struct RefreshCoordConfig {
+    /// Claim TTL applied to every L2 acquire/heartbeat call.
+    pub claim_ttl: Duration,
+    /// Cadence of background heartbeat ticks while a claim is held.
+    pub heartbeat_interval: Duration,
+    /// Maximum duration the user's refresh closure may run.
+    pub refresh_timeout: Duration,
+    /// Cadence of the background reclaim sweep (Stage 3.3).
+    pub reclaim_sweep_interval: Duration,
+    /// Sentinel events allowed inside `sentinel_window` before the
+    /// credential is escalated to `ReauthRequired` (Stage 3.2).
+    pub sentinel_threshold: u32,
+    /// Rolling window for sentinel-event counting (Stage 3.2).
+    pub sentinel_window: Duration,
+}
+
+impl Default for RefreshCoordConfig {
+    fn default() -> Self {
+        Self {
+            claim_ttl: Duration::from_secs(30),
+            heartbeat_interval: Duration::from_secs(10),
+            refresh_timeout: Duration::from_secs(8),
+            reclaim_sweep_interval: Duration::from_secs(30),
+            sentinel_threshold: 3,
+            sentinel_window: Duration::from_hours(1),
+        }
+    }
+}
+
+/// Validation errors for [`RefreshCoordConfig`].
+#[derive(Debug, thiserror::Error)]
+#[non_exhaustive]
+pub enum ConfigError {
+    /// `heartbeat_interval × 3` exceeds `claim_ttl` -- three heartbeat
+    /// ticks would not fit inside one TTL window.
+    #[error("heartbeat_interval \u{d7} 3 must be \u{2264} claim_ttl")]
+    HeartbeatTooSlow,
+    /// `refresh_timeout + 2 × heartbeat_interval` exceeds `claim_ttl` --
+    /// the holder cannot finish before its claim can expire.
+    #[error("refresh_timeout + 2 \u{d7} heartbeat_interval must be \u{2264} claim_ttl")]
+    RefreshTimeoutTooLong,
+    /// `reclaim_sweep_interval` exceeds `claim_ttl`.
+    #[error("reclaim_sweep_interval must be \u{2264} claim_ttl")]
+    ReclaimTooSlow,
+    /// A computed Duration overflowed during invariant validation. The
+    /// surfaced field name lets operators spot which knob to bound (e.g.
+    /// `heartbeat_interval × 3` or `refresh_timeout + 2 × heartbeat_interval`).
+    /// `validate()` MUST surface bad config as a typed error rather than
+    /// panicking inside the very fn meant to detect bad config.
+    #[error("config field {field} overflowed Duration during invariant check (value: {value:?})")]
+    Overflow {
+        /// Logical name of the operand that overflowed.
+        field: &'static str,
+        /// The pre-overflow operand; useful in operator messages.
+        value: Duration,
+    },
+
+    /// Metric primitive registration failed for the coordinator's series.
+    #[error("telemetry metrics error: {0}")]
+    Telemetry(#[from] nebula_metrics::MetricsError),
+}
+
+impl RefreshCoordConfig {
+    /// Verify the per- interlocking invariants.
+    ///
+    /// # Errors
+    ///
+    /// Returns `ConfigError::*` whose variant names which invariant the
+    /// configuration violates. Returns `ConfigError::Overflow` if any of
+    /// the intermediate `Duration` arithmetic (e.g. `heartbeat_interval × 3`,
+    /// `refresh_timeout + 2 × heartbeat_interval`) overflows `Duration::MAX`
+    /// -- the canonical fix is to lower the offending knob.
+    pub fn validate(&self) -> Result<(), ConfigError> {
+        // `Duration::checked_mul` and `checked_add` return `None` on
+        // overflow rather than panicking -- surface that as a typed
+        // `ConfigError::Overflow` so a user-supplied
+        // `Duration::MAX / 2`-ish value doesn't blow up the config gate.
+        let hb_x3 = self
+            .heartbeat_interval
+            .checked_mul(3)
+            .ok_or(ConfigError::Overflow {
+                field: "heartbeat_interval * 3",
+                value: self.heartbeat_interval,
+            })?;
+        if hb_x3 > self.claim_ttl {
+            return Err(ConfigError::HeartbeatTooSlow);
+        }
+        let hb_x2 = self
+            .heartbeat_interval
+            .checked_mul(2)
+            .ok_or(ConfigError::Overflow {
+                field: "heartbeat_interval * 2",
+                value: self.heartbeat_interval,
+            })?;
+        let hold_budget = self
+            .refresh_timeout
+            .checked_add(hb_x2)
+            .ok_or(ConfigError::Overflow {
+                field: "refresh_timeout + heartbeat_interval * 2",
+                value: self.refresh_timeout,
+            })?;
+        if hold_budget > self.claim_ttl {
+            return Err(ConfigError::RefreshTimeoutTooLong);
+        }
+        if self.reclaim_sweep_interval > self.claim_ttl {
+            return Err(ConfigError::ReclaimTooSlow);
+        }
+        Ok(())
+    }
+}
+
+// ──────────────────────────────────────────────────────────────────────────
+// Errors surfaced from refresh_coalesced
+// ──────────────────────────────────────────────────────────────────────────
+
+/// Failures returned by [`RefreshCoordinator::refresh_coalesced`].
+///
+/// `CoalescedByOtherReplica` is **success** for the caller (state was
+/// already fresh after another replica's refresh; just re-read state). All
+/// other variants are real errors.
+#[derive(Debug, thiserror::Error)]
+#[non_exhaustive]
+pub enum RefreshError {
+    /// Backoff retries exhausted before an L2 claim could be acquired.
+    /// Surfaced when the contender's claim keeps being heartbeat-extended.
+    #[error("contention exhausted after retries")]
+    ContentionExhausted,
+    /// Another replica's refresh succeeded while we were waiting on L2;
+    /// caller treats as success and re-reads state.
+    #[error("refresh coalesced by another replica (success \u{2014} re-read state)")]
+    CoalescedByOtherReplica,
+    /// User closure exceeded `RefreshCoordConfig::refresh_timeout`.
+    /// Distinct from `ContentionExhausted` so caller-side metrics and
+    /// retry policy can differentiate "no claim could be acquired"
+    /// from "we held the claim but the IdP call timed out".
+    #[error("refresh timeout: closure exceeded {0:?}")]
+    Timeout(Duration),
+    /// Storage repo error (e.g. DB connectivity loss).
+    #[error("storage repo error: {0}")]
+    Repo(#[from] RepoError),
+    /// Heartbeat task failure -- claim lost or repo error.
+    #[error("heartbeat error: {0}")]
+    Heartbeat(#[from] HeartbeatError),
+    /// Background heartbeat task failed mid-refresh and could not extend
+    /// the L2 claim. The user closure was aborted before issuing the IdP
+    /// POST so a stale `refresh_token_v1` cannot be sent against an
+    /// already-rotated row. Caller routes through `record_failure` and
+    /// retries with a fresh L2 acquire.
+    ///
+    /// Distinct from `Heartbeat(...)`: that variant is the surface for
+    /// the heartbeat error at construction; this variant signals the
+    /// claim was lost while a refresh closure was already running.
+    #[error(
+        "L2 claim lost during refresh \u{2014} heartbeat task failed before the IdP POST could complete"
+    )]
+    ClaimLostMidRefresh,
+    /// Configuration invariant violated at construction time.
+    #[error("config invalid: {0}")]
+    Config(#[from] ConfigError),
+}
+
+// ──────────────────────────────────────────────────────────────────────────
+// Coordinator
+// ──────────────────────────────────────────────────────────────────────────
+
+/// Two-tier credential refresh coordinator (L1 in-process + L2 cross-replica).
+pub struct RefreshCoordinator {
+    l1: L1RefreshCoalescer,
+    repo: Arc<dyn RefreshClaimRepo>,
+    replica_id: ReplicaId,
+    config: RefreshCoordConfig,
+    metrics: RefreshCoordMetrics,
+    audit_sink: Option<Arc<dyn AuditSink>>,
+}
+
+impl fmt::Debug for RefreshCoordinator {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("RefreshCoordinator")
+            .field("replica_id", &self.replica_id)
+            .field("config", &self.config)
+            .field("l1", &self.l1)
+            .field("audit_sink_present", &self.audit_sink.is_some())
+            .finish_non_exhaustive()
+    }
+}
+
+/// Result of attempting to begin a refresh for a credential -- *legacy
+/// L1-only API*.
+///
+/// Re-exported for the existing `CredentialResolver` call sites until
+/// Stage 2.3 migrates them to the closure-based
+/// [`RefreshCoordinator::refresh_coalesced`] surface. Once that migration
+/// lands, this enum (and the L1-delegate methods on `RefreshCoordinator`)
+/// can be removed.
+#[derive(Debug)]
+pub enum RefreshAttempt {
+    /// This caller won the race; perform the refresh, then call
+    /// `RefreshCoordinator::complete()` to wake waiters.
+    Winner,
+    /// Another caller is already refreshing. Await the receiver; it
+    /// resolves once the winner completes.
+    Waiter(tokio::sync::oneshot::Receiver<()>),
+}
+
+impl From<L1Attempt> for RefreshAttempt {
+    fn from(attempt: L1Attempt) -> Self {
+        match attempt {
+            L1Attempt::Winner => RefreshAttempt::Winner,
+            L1Attempt::Waiter(rx) => RefreshAttempt::Waiter(rx),
+        }
+    }
+}
+
+/// Configuration errors for the legacy concurrency knob.
+#[derive(Debug, thiserror::Error)]
+#[non_exhaustive]
+pub enum RefreshConfigError {
+    /// `max_concurrent` must be at least `1`.
+    #[error("RefreshCoordinator::max_concurrent must be >= 1, got 0")]
+    ZeroConcurrency,
+}
+
+impl RefreshCoordinator {
+    /// Maximum number of consecutive non-`ClaimLost` heartbeat
+    /// failures tolerated before the heartbeat task gives up and
+    /// cancels the in-flight refresh (sub-spec wave-4 fix).
+    ///
+    /// At three failures the worst-case latency before cancellation
+    /// is `3 × heartbeat_interval`, which is bounded by the
+    /// invariant `heartbeat_interval × 3 <= claim_ttl` -- i.e. we
+    /// never burn more than one TTL window absorbing transient
+    /// noise. Not configurable: production tuning belongs in
+    /// `RefreshCoordConfig` if a need emerges.
+    const MAX_TRANSIENT_HEARTBEAT_FAILURES: u32 = 3;
+
+    /// Construct a coordinator wired to a given `RefreshClaimRepo`.
+    ///
+    /// Metrics are bound to a fresh in-memory registry by default -- call
+    /// [`Self::with_metrics`] post-construction to thread the engine-shared
+    /// `MetricsRegistry`. Audit events are not emitted unless
+    /// [`Self::with_audit_sink`] is called.
+    ///
+    /// # Errors
+    ///
+    /// Returns the corresponding [`ConfigError`] if `config.validate()`
+    /// fails (see invariants) or metric handles cannot be bound.
+    pub fn new_with(
+        repo: Arc<dyn RefreshClaimRepo>,
+        replica_id: ReplicaId,
+        config: RefreshCoordConfig,
+    ) -> Result<Self, ConfigError> {
+        config.validate()?;
+        // Bootstrap: a fresh private registry so the coordinator is fully
+        // functional without composition. Production callers MUST follow
+        // up with `with_metrics(engine_registry)` so a scraper actually
+        // observes the series -- see `with_metrics` rustdoc.
+        let metrics = RefreshCoordMetrics::with_registry(&nebula_metrics::MetricsRegistry::new())?;
+        Ok(Self {
+            l1: L1RefreshCoalescer::new(),
+            repo,
+            replica_id,
+            config,
+            metrics,
+            audit_sink: None,
+        })
+    }
+
+    /// Replace the metric handles with ones bound to the engine-shared
+    /// `MetricsRegistry`. Call once during composition; the coordinator
+    /// emits all sub-spec series against this registry afterwards.
+    #[must_use = "builder methods must be chained or used"]
+    pub fn with_metrics(mut self, metrics: RefreshCoordMetrics) -> Self {
+        self.metrics = metrics;
+        self
+    }
+
+    /// Attach an [`AuditSink`] to receive sub-spec audit events
+    /// (`RefreshCoordClaimAcquired`, `SentinelTriggered`,
+    /// `ReauthFlagged`). Without a sink, audit emission is a no-op (the
+    /// metric / tracing surfaces still observe).
+    #[must_use = "builder methods must be chained or used"]
+    pub fn with_audit_sink(mut self, sink: Arc<dyn AuditSink>) -> Self {
+        self.audit_sink = Some(sink);
+        self
+    }
+
+    /// Borrow the pre-bound metric handles. Used by reclaim-sweep
+    /// wiring so the sweep emits the same series.
+    #[must_use]
+    pub fn metrics(&self) -> &RefreshCoordMetrics {
+        &self.metrics
+    }
+
+    /// Borrow the audit sink (`None` if not configured). Used by the
+    /// reclaim sweep to emit `RefreshCoordSentinelTriggered` /
+    /// `RefreshCoordReauthFlagged` events.
+    #[must_use]
+    pub fn audit_sink(&self) -> Option<&Arc<dyn AuditSink>> {
+        self.audit_sink.as_ref()
+    }
+
+    /// Borrow the replica identifier this coordinator was constructed
+    /// with.
+    #[must_use]
+    pub fn replica_id(&self) -> &ReplicaId {
+        &self.replica_id
+    }
+
+    /// Borrow the validated config this coordinator was constructed
+    /// with.
+    #[must_use]
+    pub fn config(&self) -> &RefreshCoordConfig {
+        &self.config
+    }
+
+    /// Borrow the underlying claim repo. Used by call sites that need
+    /// to mark the sentinel before performing the IdP POST (Stage 2.4)
+    /// and by the reclaim sweep landing in Stage 3.3.
+    pub fn repo(&self) -> &Arc<dyn RefreshClaimRepo> {
+        &self.repo
+    }
+
+    /// Acquire L1 mutex + L2 claim, run the refresh closure, release
+    /// both. Returns `Err(CoalescedByOtherReplica)` if state was already
+    /// fresh -- caller treats as success and re-reads.
+    ///
+    /// Sub-spec acquisition sequence:
+    /// 1. L1 in-process coalesce (cheap fast-path; same-process concurrent calls collapse here).
+    /// 2. L2 durable claim with backoff.
+    /// 3. Background heartbeat task -- passes `self.config.claim_ttl` to each `repo.heartbeat(token,
+    ///    ttl)` call (Stage 1 fix C2).
+    /// 4. User-supplied `do_refresh(claim)` closure.
+    /// 5. Stop heartbeat + release the claim row.
+    ///
+    /// `needs_refresh_after_backoff` is consulted by the L2 backoff loop
+    /// per sub-spec after the post-`Contended` sleep: if the
+    /// predicate returns `false` the credential was refreshed by another
+    /// replica while this caller was waiting and we short-circuit with
+    /// [`RefreshError::CoalescedByOtherReplica`]. Callers that don't
+    /// have a state-check available pass `|_| async { true }` -- that
+    /// preserves the legacy "always retry" behavior at the cost of
+    /// occasionally running the refresh closure on a freshly-refreshed
+    /// credential.
+    ///
+    /// # Errors
+    ///
+    /// See [`RefreshError`]. `CoalescedByOtherReplica` is success-with-side-effect:
+    /// another replica refreshed while we were waiting. Caller should
+    /// re-read the credential state and proceed.
+    ///
+    /// # Cancel-safety
+    ///
+    /// The `do_refresh` future MUST be cancel-safe: dropping it mid-`await`
+    /// must not leak resources or leave external state in an inconsistent
+    /// shape. The coordinator drops the future without notice if a heartbeat
+    /// failure mid-refresh requires aborting before the IdP POST issues
+    /// (returning [`RefreshError::ClaimLostMidRefresh`]).
+    ///
+    /// Standard async HTTP clients (`reqwest`, `hyper`) satisfy this. Closures
+    /// that await a `tokio::task::JoinHandle` from `spawn_blocking` without a
+    /// cleanup path do not -- the blocking task continues running after the
+    /// `JoinHandle` is dropped, and any state it produces is silently
+    /// discarded.
+    ///
+    /// The coordinator itself is cancel-safe: dropping the
+    /// `refresh_coalesced` future at any await point (`tokio::time::timeout`,
+    /// `tokio::select!`, `JoinHandle::abort`) cancels the heartbeat
+    /// task and best-effort releases the L2 claim row (spawned, fire-and-
+    /// forget; release errors are logged at WARN). The synchronous-release
+    /// contract on the success path is preserved -- when `refresh_coalesced`
+    /// returns `Ok(...)`, the L2 row has been awaited-released before the
+    /// function returns.
+    #[tracing::instrument(
+        name = "credential.refresh.coordinate",
+        skip(self, needs_refresh_after_backoff, do_refresh),
+        fields(
+            credential_id = %credential_id,
+            replica_id = %self.replica_id,
+            tier = tracing::field::Empty,
+        ),
+    )]
+    pub async fn refresh_coalesced<F, Fut, T, P, PFut>(
+        &self,
+        credential_id: &CredentialId,
+        needs_refresh_after_backoff: P,
+        do_refresh: F,
+    ) -> Result<T, RefreshError>
+    where
+        // Explicit `Send` bounds (review I2): the inner futures cross
+        // task boundaries because `do_refresh` runs under
+        // `tokio::time::timeout` and the predicate is awaited from the
+        // spawn'd backoff loop. Without these bounds a `!Send` body
+        // (e.g. one that captures an `Rc<...>`) compiles cleanly here
+        // and surfaces an obscure auto-trait error at the call site.
+        // Locking the contract on the trait bound moves the diagnostic
+        // back to the user closure.
+        F: FnOnce(RefreshClaim) -> Fut + Send,
+        Fut: Future<Output = Result<T, RefreshError>> + Send,
+        T: Send,
+        P: Fn(&CredentialId) -> PFut + Sync,
+        PFut: Future<Output = bool> + Send,
+    {
+        // L1: in-process coalescing.
+        //
+        // The L1 layer is keyed by string, so we hash on the typed id's
+        // canonical form. `try_refresh` returns Winner for the first
+        // caller and Waiter (with a oneshot::Receiver) for every other
+        // concurrent caller in the same process. Waiters await the
+        // Winner's `complete()` call, then surface
+        // `CoalescedByOtherReplica` so the caller re-reads state -- by
+        // construction the Winner has already released the L2 claim and
+        // committed the refreshed state to storage.
+        let cred_str = credential_id.to_string();
+        match self.l1.try_refresh(&cred_str) {
+            super::l1::RefreshAttempt::Winner => {
+                // NOTE: do NOT record `tier="l2"` here -- the L2 path can
+                // still produce `CoalescedByOtherReplica` via the
+                // post-backoff recheck in
+                // `try_acquire_l2_with_backoff`. Recording the tier
+                // prematurely makes operators see "l2 acquired" when the
+                // actual outcome was "l2 coalesced" (review I1).
+                // The closed set `{l1, l2_acquired, l2_coalesced}` is
+                // recorded at the actual outcome sites below.
+                // (Fall through to acquire L2 + run the user closure.)
+            },
+            super::l1::RefreshAttempt::Waiter(rx) => {
+                tracing::Span::current().record("tier", "l1");
+                self.metrics.coalesced_l1.inc();
+                // Wait for the Winner to finish. If the receiver errors
+                // (Winner panicked / dropped without complete()) we still
+                // re-read state -- pessimistic safety.
+                let _ = rx.await;
+                return Err(RefreshError::CoalescedByOtherReplica);
+            },
+        }
+
+        // Make sure `complete()` runs even on early return / panic so
+        // future callers can re-acquire the L1 slot.
+        let credential_id_for_guard = cred_str.clone();
+        let l1 = &self.l1;
+        let _l1_complete = scopeguard::guard((), move |()| {
+            l1.complete(&credential_id_for_guard);
+        });
+
+        // Global rate-limit gate (audit B6 / wave-2 regression).
+        //
+        // Wave-2 introduced this typed entry point but silently bypassed
+        // the L1 global concurrency semaphore (`refresh_semaphore`,
+        // default 32 permits). Per-credential L1 coalescing alone does
+        // not bound the case where many *distinct* credentials expire
+        // near-simultaneously -- e.g. on a daily TTL boundary or after
+        // a replica restart with stale tokens -- and a 200-credential
+        // expiry burst would issue 200 concurrent IdP POSTs, recreating
+        // the cascading-429 / refresh-storm pattern the cap is meant to
+        // prevent. Only the legacy `String`-id path
+        // (`resolver.rs::refresh_via_l1_only`) consumed permits, so
+        // typed callers were unprotected.
+        //
+        // Acquired AFTER `try_refresh` (Winner-only -- Waiters already
+        // park on the oneshot above and do not need a permit) and BEFORE
+        // L2 backoff so the bound covers the entire IdP POST window.
+        // The order also keeps `_l1_complete` declared first so its
+        // guard fires on every cancel/Drop path even if `acquire_permit`
+        // itself is cancelled (its `await` is cancel-safe per
+        // `L1RefreshCoalescer::acquire_permit` rustdoc -- dropping the
+        // future does not consume a permit).
+        //
+        // RAII: `_permit` holds an `OwnedSemaphorePermit` until end of
+        // function (after explicit synchronous release on the success
+        // path; after `l2_teardown` fires on every other path), so the
+        // permit is released on every exit including Drop and panic.
+        let _permit = self.l1.acquire_permit().await;
+
+        // L2: durable claim with backoff.
+        let claim = self
+            .try_acquire_l2_with_backoff(credential_id, &needs_refresh_after_backoff)
+            .await?;
+
+        // Sub-spec -- record the claim acquisition once we know we own
+        // the L2 row. `acquired` counter, audit event, and start of the
+        // hold-duration measurement happen here so they are paired
+        // with the matching `release` site below.
+        //
+        // Span tier (review I1) -- record `l2_acquired` at the outcome
+        // site so operators distinguish from the `l2_coalesced` path
+        // (post-backoff recheck), which is recorded inside
+        // `try_acquire_l2_with_backoff` below.
+        tracing::Span::current().record("tier", "l2_acquired");
+        self.metrics.claims_acquired.inc();
+        emit_claim_acquired(
+            self.audit_sink.as_deref(),
+            credential_id,
+            self.replica_id.as_str(),
+            self.config.claim_ttl.as_secs(),
+        );
+        let hold_start = Instant::now();
+
+        // Cancellation token shared between heartbeat task and the user
+        // closure. The heartbeat task fires `cancel()` on Err so
+        // `do_refresh` can abort before the IdP POST runs against an
+        // already-rotated row (n8n #13088 lineage).
+        let cancel = CancellationToken::new();
+        let hb_cancel = cancel.clone();
+
+        // Heartbeat task in background.
+        let hb_task = self.spawn_heartbeat(claim.token.clone(), hb_cancel, *credential_id);
+
+        // Cancel-safety guard (review C1 + wave-5, sub-spec). Fires
+        // on EVERY exit path that does NOT explicitly defuse it: panic
+        // unwind, error early-return, AND Drop (caller cancels the
+        // outer future via `tokio::time::timeout`, `tokio::select!`, or
+        // `JoinHandle::abort`). Without Drop coverage, a caller that
+        // wraps `refresh_coalesced` in its own timeout / select! /
+        // spawn-and-drop would leak the heartbeat task (extending L2
+        // expiry forever) and skip `release()` -- the same "stuck claim"
+        // failure mode the panic-only guard was designed to prevent,
+        // just triggered by `Drop` instead of unwind.
+        //
+        // The success path defuses this guard with
+        // `scopeguard::ScopeGuard::into_inner(...)` BEFORE running its
+        // synchronous release -- that preserves the synchronous-release
+        // contract on `Ok(...)` while making every other exit do
+        // best-effort spawned cleanup. `release()` is idempotent and
+        // the spawned task is detached because Drop is synchronous; we
+        // cannot `.await` here.
+        //
+        // Heartbeat shutdown order mirrors the success path (below):
+        // `cancel.cancel()` first so the heartbeat exits through its
+        // `cancelled()` arm cleanly, then `abort()` as a belt-and-suspenders
+        // guarantee. `CancellationToken` is reference-counted so dropping
+        // it does NOT auto-cancel -- both paths must call `.cancel()`
+        // explicitly to keep release semantics symmetric.
+        let token_teardown = claim.token.clone();
+        let repo_teardown = Arc::clone(&self.repo);
+        let hb_cancel_teardown = cancel.clone();
+        let hb_task_teardown = hb_task.abort_handle();
+        let hold_duration_teardown = self.metrics.hold_duration.clone();
+        let l2_teardown = scopeguard::guard((), move |()| {
+            hb_cancel_teardown.cancel();
+            hb_task_teardown.abort();
+            // Hold-time histogram is recorded on every exit path --
+            // observe it before the spawn so the histogram never drops
+            // a sample regardless of which path triggered teardown.
+            hold_duration_teardown.observe(hold_start.elapsed().as_secs_f64());
+            tokio::spawn(async move {
+                if let Err(e) = repo_teardown.release(token_teardown).await {
+                    tracing::warn!(?e, "L2 claim release on drop/unwind failed");
+                }
+            });
+        });
+
+        // Keep the token for the normal-exit release; `do_refresh`
+        // takes the full `RefreshClaim` (which it may inspect for
+        // `expires_at`), so we hand it a clone and retain `token` here.
+        let token_for_release = claim.token.clone();
+
+        // Run user's refresh closure under `refresh_timeout`.
+        // The timeout is shorter than the claim TTL by construction, so
+        // the heartbeat keeps the L2 row alive while the closure runs.
+        // Wrap in `select!` over `cancel.cancelled()` so a heartbeat
+        // failure mid-refresh aborts the closure BEFORE it issues the
+        // IdP POST -- sub-spec invariant.
+        //
+        // Bias order MUST poll `do_refresh_fut` first: if the future
+        // resolves `Ok(...)` and the heartbeat task fires
+        // `cancel.cancel()` in the same wake-cycle, biased-cancel-first
+        // would silently drop the successful result and route through
+        // `record_failure`, which would reissue an IdP POST against a
+        // stale `refresh_token_v1` -- re-introducing the n8n #13088
+        // refresh-storm pattern P2 was designed to prevent. With
+        // `do_refresh_fut` polled first, a ready future deterministically
+        // wins; cancel only fires while the refresh future is still
+        // suspended. The `select!` is wrapped INSIDE `timeout(...)` so
+        // the timeout still bounds the overall wait.
+        let timeout = self.config.refresh_timeout;
+        let do_refresh_fut = do_refresh(claim);
+        let result = tokio::time::timeout(timeout, async {
+            tokio::select! {
+                biased;
+                r = do_refresh_fut => r,
+                () = cancel.cancelled() => Err(RefreshError::ClaimLostMidRefresh),
+            }
+        })
+        .await
+        .map_err(|_elapsed| RefreshError::Timeout(timeout))
+        .and_then(std::convert::identity);
+
+        // Normal-exit release (review I1 + wave-5, sub-spec). We
+        // DO NOT propagate release errors -- propagating them with `?`
+        // would mask a successful refresh: caller would observe
+        // `RefreshError::Repo(...)`, route to `record_failure`, then
+        // retry -> ANOTHER IdP POST -> invalidates the just-issued
+        // refresh token (n8n #13088 spec lineage). Log at warn level
+        // instead.
+        //
+        // `into_inner` defuses the teardown guard BEFORE the explicit
+        // cleanup runs. Without this, the guard would also fire when
+        // `l2_teardown` goes out of scope at end of function, racing
+        // the synchronous release below (double-release; second one
+        // sees a missing row -> spurious WARN log). Defusing first
+        // preserves the synchronous-release contract: when
+        // `refresh_coalesced` returns `Ok(...)`, the L2 row has been
+        // awaited-released before the function returns.
+        scopeguard::ScopeGuard::into_inner(l2_teardown);
+        // Cancel the heartbeat token before aborting so the task exits
+        // through its `cancelled()` arm rather than racing the abort.
+        cancel.cancel();
+        hb_task.abort();
+        // Sub-spec -- observe the hold duration on the normal-exit
+        // release path. Symmetric with the teardown guard above.
+        self.metrics
+            .hold_duration
+            .observe(hold_start.elapsed().as_secs_f64());
+        if let Err(e) = self.repo.release(token_for_release).await {
+            tracing::warn!(?e, "L2 claim release after successful refresh failed");
+        }
+
+        result
+    }
+
+    /// L2 acquisition retry loop per sub-spec.
+    ///
+    /// On `Contended` we sleep until the contender's claim is expected
+    /// to expire (capped + jitter) then consult
+    /// `needs_refresh_after_backoff(credential_id)`. If the predicate
+    /// returns `false` we surface
+    /// [`RefreshError::CoalescedByOtherReplica`] -- another replica
+    /// completed the refresh while we were waiting, and the caller
+    /// should re-read state from storage. Otherwise we retry
+    /// `try_claim` until we win the claim or exhaust attempts.
+    async fn try_acquire_l2_with_backoff<P, PFut>(
+        &self,
+        credential_id: &CredentialId,
+        needs_refresh_after_backoff: &P,
+    ) -> Result<RefreshClaim, RefreshError>
+    where
+        // Mirror the `Send`/`Sync` bounds on `refresh_coalesced` so the
+        // helper's auto-trait inference does not silently relax the
+        // public contract.
+        P: Fn(&CredentialId) -> PFut + Sync,
+        PFut: Future<Output = bool> + Send,
+    {
+        const MAX_ATTEMPTS: usize = 5;
+        for attempt in 0..MAX_ATTEMPTS {
+            // Sub-spec per-attempt tracing span: `attempt` and
+            // `credential_id` so operators correlate contention storms
+            // across replicas.
+            let span = tracing::info_span!(
+                "credential.refresh.claim.acquire",
+                credential_id = %credential_id,
+                replica_id = %self.replica_id,
+                attempt = attempt,
+            );
+            let outcome = async {
+                self.repo
+                    .try_claim(credential_id, &self.replica_id, self.config.claim_ttl)
+                    .await
+            }
+            .instrument(span)
+            .await?;
+            match outcome {
+                ClaimAttempt::Acquired(claim) => return Ok(claim),
+                ClaimAttempt::Contended {
+                    existing_expires_at,
+                } => {
+                    // Sub-spec -- bump the contended counter for every
+                    // try_claim that returned Contended, regardless of
+                    // whether the post-backoff recheck eventually
+                    // short-circuits.
+                    self.metrics.claims_contended.inc();
+                    // Sleep until the contender's claim expires (capped
+                    // at 5s so we don't sleep forever if their TTL is
+                    // somehow much longer than ours), plus a jitter to
+                    // de-correlate retries across replicas.
+                    let now = chrono::Utc::now();
+                    let cap = now + chrono::Duration::seconds(5);
+                    let wait_until = existing_expires_at.min(cap);
+                    let delay = (wait_until - now)
+                        .to_std()
+                        .unwrap_or(Duration::from_millis(200));
+                    tokio::time::sleep(delay + jitter_ms(100)).await;
+                    // CRITICAL: post-backoff state recheck per sub-spec. If
+                    // the contender finished the refresh while we slept,
+                    // the credential is now fresh -- short-circuit with
+                    // CoalescedByOtherReplica so the caller re-reads
+                    // state instead of running another IdP POST. Without
+                    // this check, two replicas racing through L2 each
+                    // run the closure (one wins try_claim now that the
+                    // contender's row is gone), invalidating any
+                    // refresh_token rotation the contender just
+                    // committed (n8n #13088 lineage).
+                    if !needs_refresh_after_backoff(credential_id).await {
+                        // Sub-spec -- L2 coalesce: another replica
+                        // refreshed while we waited.
+                        //
+                        // Span tier (review I1) -- record `l2_coalesced`
+                        // at the outcome site. We are now outside the
+                        // per-attempt `instrument(span)` block (which
+                        // wrapped only the `try_claim` future), so
+                        // `Span::current()` resolves to the parent
+                        // `credential.refresh.coordinate` span -- the
+                        // intended target. The closed set
+                        // `{l1, l2_acquired, l2_coalesced}` is
+                        // documented in OBSERVABILITY.md.
+                        tracing::Span::current().record("tier", "l2_coalesced");
+                        self.metrics.coalesced_l2.inc();
+                        return Err(RefreshError::CoalescedByOtherReplica);
+                    }
+                },
+            }
+        }
+        // Sub-spec -- every retry exhausted without acquiring the L2
+        // row. `claims_total{outcome=exhausted} > 0` is a real production
+        // signal worth alerting on.
+        self.metrics.claims_exhausted.inc();
+        Err(RefreshError::ContentionExhausted)
+    }
+
+    /// Spawn the background heartbeat task that refreshes the L2 claim
+    /// TTL on a fixed interval. Per Stage 1 fix C2 the trait's
+    /// `heartbeat(token, ttl)` takes the same TTL passed to
+    /// `try_claim`, so the invariants
+    /// (`heartbeat_interval × 3 < claim_ttl`,
+    /// `reclaim_sweep_interval <= claim_ttl`) hold across heartbeats.
+    ///
+    /// Exits and signals cancellation via the supplied
+    /// [`CancellationToken`] in two cases:
+    ///
+    /// 1. **Claim lost** (`HeartbeatError::ClaimLost`): a different replica reclaimed the row
+    ///    (generation bumped or row deleted). Cancellation fires immediately so the in-flight
+    ///    refresh aborts BEFORE the IdP POST. Without this signal the closure would press on with a
+    ///    stale `refresh_token_v1` and invalidate the just-rotated row (n8n #13088 lineage).
+    ///
+    /// 2. **Transient errors past budget**: any non-`ClaimLost` heartbeat error (e.g. transient
+    ///    backend hiccup wrapped in `HeartbeatError::Repo`) retries up to
+    ///    [`Self::MAX_TRANSIENT_HEARTBEAT_FAILURES`] times. Single transient hiccups are absorbed
+    ///    silently so storage backpressure does not amplify into refresh storms. After the budget
+    ///    is exhausted, cancellation fires.
+    fn spawn_heartbeat(
+        &self,
+        token: ClaimToken,
+        cancel: CancellationToken,
+        credential_id: CredentialId,
+    ) -> tokio::task::JoinHandle<()> {
+        let repo = Arc::clone(&self.repo);
+        let interval = self.config.heartbeat_interval;
+        let ttl = self.config.claim_ttl;
+        let replica_id = self.replica_id.as_str().to_string();
+        tokio::spawn(async move {
+            let mut ticker = tokio::time::interval(interval);
+            // Avoid heartbeat amplification under storage backpressure:
+            // if a heartbeat call exceeds `interval`, drop missed ticks
+            // rather than firing them back-to-back when the call returns.
+            ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
+            // Burn the initial immediate tick -- the claim was just
+            // acquired and already has a fresh `expires_at`.
+            ticker.tick().await;
+            // Transient-failure budget per sub-spec (wave-4 fix).
+            // Resets on every successful heartbeat so a long-running
+            // refresh can absorb intermittent backend noise without
+            // cancelling. Only `HeartbeatError::ClaimLost` is treated
+            // as immediate-cancel -- that is the unambiguous "your
+            // claim is gone" signal.
+            let mut transient_failures: u32 = 0;
+            loop {
+                tokio::select! {
+                    biased;
+                    () = cancel.cancelled() => {
+                        // Caller wound up its refresh -- heartbeat exits
+                        // cleanly, no signal to fire.
+                        break;
+                    }
+                    _ = ticker.tick() => {
+                        match repo.heartbeat(&token, ttl).await {
+                            Ok(()) => {
+                                // Reset the transient-failure budget on
+                                // every success so a long refresh can
+                                // absorb intermittent noise.
+                                transient_failures = 0;
+                            }
+                            Err(HeartbeatError::ClaimLost) => {
+                                // ERROR-level: claim loss is the
+                                // unambiguous "another replica reclaimed
+                                // the row" signal. Promoting from WARN
+                                // keeps it distinguishable from
+                                // transient retry noise on dashboards
+                                // filtering on level.
+                                tracing::error!(
+                                    %credential_id,
+                                    replica_id = %replica_id,
+                                    "credential refresh heartbeat lost claim; cancelling concurrent \
+                                     do_refresh before it issues the IdP POST"
+                                );
+                                // Fire cancel BEFORE breaking so the user
+                                // closure sees `cancelled()` and aborts
+                                // the IdP call.
+                                cancel.cancel();
+                                break;
+                            }
+                            Err(HeartbeatError::Repo(repo_err)) => {
+                                // Variant-explicit on purpose:
+                                // `HeartbeatError` is NOT `#[non_exhaustive]`,
+                                // so a wildcard `Err(other)` would silently
+                                // bucket any future variant (e.g.
+                                // `Unauthorized`, `Throttled`) as transient.
+                                // Matching `Repo(_)` explicitly forces a
+                                // compiler error when a new variant is
+                                // added so the next maintainer makes a
+                                // per-variant policy decision rather than
+                                // inheriting "treat as transient" by accident.
+                                transient_failures += 1;
+                                if transient_failures >= Self::MAX_TRANSIENT_HEARTBEAT_FAILURES {
+                                    tracing::error!(
+                                        error = ?repo_err,
+                                        %credential_id,
+                                        replica_id = %replica_id,
+                                        attempts = transient_failures,
+                                        max_attempts = Self::MAX_TRANSIENT_HEARTBEAT_FAILURES,
+                                        "credential refresh heartbeat exceeded transient-failure \
+                                         budget; cancelling concurrent do_refresh before it issues \
+                                         the IdP POST"
+                                    );
+                                    cancel.cancel();
+                                    break;
+                                }
+                                // Log at WARN -- single hiccups are
+                                // absorbed silently from a level-filter
+                                // perspective. Operators can still see
+                                // them on noisy-log dashboards.
+                                tracing::warn!(
+                                    error = ?repo_err,
+                                    %credential_id,
+                                    replica_id = %replica_id,
+                                    attempt = transient_failures,
+                                    max_attempts = Self::MAX_TRANSIENT_HEARTBEAT_FAILURES,
+                                    "credential refresh heartbeat transient error; retrying \
+                                     within budget"
+                                );
+                                // Continue -- next ticker tick will retry.
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    }
+
+    // ──────────────────────────────────────────────────────────────────
+    // Legacy L1 surface -- kept until Stage 2.3 migrates the resolver
+    // to `refresh_coalesced`. Each method delegates to the inner
+    // L1 coalescer.
+    // ──────────────────────────────────────────────────────────────────
+
+    /// **Legacy.** Begin a refresh attempt against the L1 coalescer.
+    /// Stage 2.3 deletes this in favour of [`Self::refresh_coalesced`].
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn try_refresh(&self, credential_id: &str) -> RefreshAttempt {
+        self.l1.try_refresh(credential_id).into()
+    }
+
+    /// **Legacy.** Mark the L1 in-flight slot complete. Stage 2.3
+    /// deletes this in favour of [`Self::refresh_coalesced`].
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn complete(&self, credential_id: &str) {
+        self.l1.complete(credential_id);
+    }
+
+    /// **Legacy.** Number of credentials currently being refreshed in
+    /// the L1 layer.
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn in_flight_count(&self) -> usize {
+        self.l1.in_flight_count()
+    }
+
+    /// **Legacy.** Acquire a permit from the L1 concurrency limiter.
+    ///
+    /// Only the standalone permit-grab API is deprecated; the underlying
+    /// global concurrency semaphore is **not** going away. The typed-path
+    /// `Self::refresh_coalesced` acquires from the same semaphore
+    /// internally (Winner-only, RAII-scoped), so callers migrating off
+    /// the legacy `String`-id surface inherit the rate-limit defense
+    /// without any explicit wiring. See audit B6 / wave-2 regression.
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub async fn acquire_permit(&self) -> tokio::sync::OwnedSemaphorePermit {
+        self.l1.acquire_permit().await
+    }
+
+    /// **Legacy.** Available permits in the L1 concurrency limiter.
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn available_permits(&self) -> usize {
+        self.l1.available_permits()
+    }
+
+    /// **Legacy.** Record a refresh failure for the L1 circuit breaker.
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn record_failure(&self, credential_id: &str) {
+        self.l1.record_failure(credential_id);
+    }
+
+    /// **Legacy.** Record a refresh success for the L1 circuit breaker.
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn record_success(&self, credential_id: &str) {
+        self.l1.record_success(credential_id);
+    }
+
+    /// **Legacy.** Whether the L1 per-credential circuit breaker is
+    /// open.
+    #[deprecated(
+        since = "0.1.0",
+        note = "use refresh_coalesced; remove when typed CredentialId migration completes"
+    )]
+    pub fn is_circuit_open(&self, credential_id: &str) -> bool {
+        self.l1.is_circuit_open(credential_id)
+    }
+}
+
+// `Default` deliberately not implemented: the only constructor without an
+// explicit repo arg would need `InMemoryRefreshClaimRepo` from `nebula-storage`,
+// which is outside this crate's dep graph. Callers construct via
+// `RefreshCoordinator::new_with(repo, replica_id, config)`.
+
+// ──────────────────────────────────────────────────────────────────────────
+// Helpers
+// ──────────────────────────────────────────────────────────────────────────
+
+fn jitter_ms(max_ms: u64) -> Duration {
+    if max_ms == 0 {
+        return Duration::ZERO;
+    }
+    let amount = rand::random_range(0..max_ms);
+    Duration::from_millis(amount)
+}
