+++ crates/credential/src/runtime/refresh/mod.rs
@@ -7,15 +7,15 @@
 //! `L1RefreshCoalescer` so the rename is atomic and existing callers
 //! (`CredentialResolver`) keep compiling. Stage 2.2 replaces the wrapper
 //! with the real two-tier acquisition (`refresh_coalesced(...)` acquires
-//! L1 mutex first, then a durable L2 claim via `RefreshClaimRepo`).
+//! L1 mutex first, then a durable L2 claim via `RefreshClaimStore`).
 //!
 //! # Layering
 //!
 //! - **L1 (in-process):** `L1RefreshCoalescer` (`l1.rs`, private). Coalesces concurrent refreshes
 //!   inside the same replica process via per-credential `oneshot` waiters and a global concurrency
 //!   semaphore. Includes a per-credential `nebula_resilience::CircuitBreaker`.
-//! - **L2 (cross-replica):** `RefreshClaimRepo` (in `nebula-storage`). CAS-based claim with TTL +
-//!   heartbeat. Lands in Stage 2.2.
+//! - **L2 (cross-replica):** `RefreshClaimStore` (in `nebula-storage-port`). CAS-based claim with
+//!   TTL + heartbeat.
 
 mod audit;
 mod coordinator;
@@ -23,9 +23,6 @@ mod l1;
 mod metrics;
 mod reclaim;
 mod sentinel;
-
-#[cfg(test)]
-mod test_fixtures;
 
 pub use coordinator::{
     ConfigError, RefreshAttempt, RefreshConfigError, RefreshCoordConfig, RefreshCoordinator,
