--- crates/engine/src/credential/refresh/reclaim.rs
@@ -1,601 +1,0 @@
-//! Background reclaim-sweep task. Parallel to control-queue reclaim.
-//!
-//! Per sub-spec
-//! `docs/INTEGRATION_MODEL.md`
-//! +.
-//!
-//! On a fixed cadence the task calls `RefreshClaimRepo::reclaim_stuck`,
-//! atomically sweeping every claim row whose `expires_at` is in the
-//! past. For each row the sweep returns paired with its observed
-//! sentinel state:
-//!
-//! - `SentinelState::Normal` — the holder simply timed out without ever entering the IdP POST. The
-//!   row is already deleted by `reclaim_stuck`; nothing more to do.
-//! - `SentinelState::RefreshInFlight` — the holder crashed mid-refresh. The sweep routes the event
-//!   to [`SentinelTrigger::on_sentinel_detected`] which records it in `credential_sentinel_events`
-//!   and consults the rolling-window count. When the threshold is reached the sweep publishes
-//!   `CredentialEvent::ReauthRequired { credential_id, reason: SentinelRepeated }` so downstream
-//!   consumers (UI, monitoring, resource pools) react immediately.
-//!
-//! Two sweeps observing the same expired row are prevented at the
-//! storage layer — `reclaim_stuck` issues one atomic `DELETE ... RETURNING`
-//! so each row appears in exactly one sweeper's result.
-
-use std::{sync::Arc, time::Duration};
-
-use nebula_credential::{CredentialEvent, resolve::ReauthReason};
-use nebula_eventbus::EventBus;
-use nebula_storage::credential::{AuditSink, RefreshClaimRepo, SentinelState};
-use tracing::Instrument;
-
-use super::{
-    audit::{emit_reauth_flagged, emit_sentinel_triggered},
-    coordinator::RefreshCoordinator,
-    metrics::RefreshCoordMetrics,
-    sentinel::{SentinelDecision, SentinelTrigger},
-};
-
-/// Handle for the background reclaim sweep task.
-///
-/// Holding the handle keeps the task alive; dropping or calling
-/// [`Self::abort`] cancels it. The task itself loops forever — the
-/// handle is the only path to graceful shutdown.
-pub struct ReclaimSweepHandle {
-    handle: tokio::task::JoinHandle<()>,
-}
-
-impl std::fmt::Debug for ReclaimSweepHandle {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ReclaimSweepHandle")
-            .field("is_finished", &self.handle.is_finished())
-            .finish()
-    }
-}
-
-impl ReclaimSweepHandle {
-    /// Spawn the reclaim sweep task wired to a [`RefreshCoordinator`].
-    ///
-    /// The cadence and the underlying [`RefreshClaimRepo`] are derived
-    /// from `coord` so the invariant
-    /// `reclaim_sweep_interval ≤ claim_ttl` (validated at coordinator
-    /// construction) reaches the sweep task by construction — no
-    /// freestanding `Duration` argument that could drift from the
-    /// validated config (review feedback I3).
-    ///
-    /// The task wakes every `coord.config().reclaim_sweep_interval`,
-    /// calls `repo.reclaim_stuck()`, routes any `RefreshInFlight`
-    /// claims to `sentinel.on_sentinel_detected`, and emits
-    /// `CredentialEvent::ReauthRequired` on the supplied event bus
-    /// when the threshold is exceeded.
-    ///
-    /// `event_bus` is optional: in tests / desktop mode without an
-    /// event bus the threshold-exceed path still records the event in
-    /// `credential_sentinel_events` and logs a `tracing::warn!` line —
-    /// only the cross-replica fan-out is skipped.
-    pub fn spawn(
-        coord: Arc<RefreshCoordinator>,
-        sentinel: Arc<SentinelTrigger>,
-        event_bus: Option<Arc<EventBus<CredentialEvent>>>,
-    ) -> Self {
-        let cadence = coord.config().reclaim_sweep_interval;
-        let repo = Arc::clone(coord.repo());
-        // Sub-spec — sweep emits metrics + audit events through the
-        // same handles wired into the coordinator so a single PromQL /
-        // sink view aggregates both refresh sites.
-        let metrics = coord.metrics().clone();
-        let audit_sink = coord.audit_sink().cloned();
-        let handle = tokio::spawn(async move {
-            sweep_loop(repo, sentinel, cadence, event_bus, metrics, audit_sink).await;
-        });
-        Self { handle }
-    }
-
-    /// Abort the running sweep task. Safe to call multiple times.
-    pub fn abort(&self) {
-        self.handle.abort();
-    }
-
-    /// Whether the underlying task has finished (e.g. via abort).
-    #[must_use]
-    pub fn is_finished(&self) -> bool {
-        self.handle.is_finished()
-    }
-}
-
-impl Drop for ReclaimSweepHandle {
-    fn drop(&mut self) {
-        // Cancel the spawned task so the sweep does not outlive the
-        // engine that started it.
-        self.handle.abort();
-    }
-}
-
-async fn sweep_loop(
-    repo: Arc<dyn RefreshClaimRepo>,
-    sentinel: Arc<SentinelTrigger>,
-    cadence: Duration,
-    event_bus: Option<Arc<EventBus<CredentialEvent>>>,
-    metrics: RefreshCoordMetrics,
-    audit_sink: Option<Arc<dyn AuditSink>>,
-) {
-    let mut ticker = tokio::time::interval(cadence);
-    // Avoid back-to-back sweeps after a long storage stall.
-    ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-    // Burn the leading immediate tick so the first real sweep happens
-    // after one cadence — mirrors the heartbeat task pattern in
-    // `coordinator.rs`.
-    ticker.tick().await;
-    loop {
-        ticker.tick().await;
-        if let Err(e) = run_one_sweep(
-            &repo,
-            &sentinel,
-            event_bus.as_ref(),
-            &metrics,
-            audit_sink.as_deref(),
-        )
-        .await
-        {
-            // run_one_sweep already logs per-row errors; this catches
-            // the top-level reclaim_stuck failure.
-            tracing::warn!(?e, "credential refresh reclaim sweep failed");
-        }
-    }
-}
-
-async fn run_one_sweep(
-    repo: &Arc<dyn RefreshClaimRepo>,
-    sentinel: &Arc<SentinelTrigger>,
-    event_bus: Option<&Arc<EventBus<CredentialEvent>>>,
-    metrics: &RefreshCoordMetrics,
-    audit_sink: Option<&dyn AuditSink>,
-) -> Result<(), nebula_storage::credential::RepoError> {
-    let stuck = repo.reclaim_stuck().await?;
-    // Sub-spec — increment reclaim sweep counter once per sweep,
-    // labeled by whether work was found. `no_work` is the steady state
-    // for healthy systems; `reclaimed` rising is a crashed-runner signal.
-    if stuck.is_empty() {
-        metrics.reclaim_no_work.inc();
-    } else {
-        metrics.reclaim_reclaimed.inc();
-    }
-    for reclaimed in stuck {
-        if reclaimed.sentinel != SentinelState::RefreshInFlight {
-            // Normal-path expiry — claim was reclaimed, nothing more to
-            // do (no mid-refresh crash to record).
-            continue;
-        }
-        // Sub-spec — per-row span; an operator can grep
-        // `credential.refresh.sentinel.detected` for crashed-mid-refresh
-        // events without wading through normal-expiry rows. Use
-        // `.instrument(...)` rather than `.entered()` so the span
-        // boundary respects `Send` for the spawned sweep task.
-        let detect_span = tracing::info_span!(
-            "credential.refresh.sentinel.detected",
-            credential_id = %reclaimed.credential_id,
-            crashed_holder = %reclaimed.previous_holder,
-            generation = reclaimed.previous_generation,
-        );
-        let decision = match async {
-            sentinel
-                .on_sentinel_detected(
-                    &reclaimed.credential_id,
-                    &reclaimed.previous_holder,
-                    reclaimed.previous_generation,
-                )
-                .await
-        }
-        .instrument(detect_span)
-        .await
-        {
-            Ok(d) => d,
-            Err(e) => {
-                tracing::warn!(
-                    cred = %reclaimed.credential_id,
-                    ?e,
-                    "sentinel decision failed; reclaim sweep continues"
-                );
-                continue;
-            },
-        };
-
-        match decision {
-            SentinelDecision::Recoverable { event_count } => {
-                tracing::info!(
-                    cred = %reclaimed.credential_id,
-                    event_count,
-                    "sentinel recoverable — credential refresh will retry"
-                );
-                // Sub-spec — record the event (below threshold).
-                metrics.sentinel_recorded.inc();
-                emit_sentinel_triggered(audit_sink, &reclaimed.credential_id, event_count);
-            },
-            SentinelDecision::EscalateToReauth {
-                event_count,
-                window_secs,
-            } => {
-                tracing::warn!(
-                    cred = %reclaimed.credential_id,
-                    event_count,
-                    window_secs,
-                    "sentinel threshold exceeded — escalating to ReauthRequired"
-                );
-                // Sub-spec — bump both the recorded counter (every
-                // detection counts) and the reauth_triggered counter
-                // (the escalation transition itself).
-                metrics.sentinel_recorded.inc();
-                metrics.sentinel_reauth_triggered.inc();
-                emit_sentinel_triggered(audit_sink, &reclaimed.credential_id, event_count);
-                emit_reauth_flagged(audit_sink, &reclaimed.credential_id, "sentinel_repeated");
-                if let Some(bus) = event_bus {
-                    let event = CredentialEvent::ReauthRequired {
-                        credential_id: reclaimed.credential_id,
-                        reason: ReauthReason::SentinelRepeated {
-                            event_count,
-                            window_secs,
-                        },
-                    };
-                    let outcome = bus.emit(event);
-                    if !matches!(outcome, nebula_eventbus::PublishOutcome::Sent) {
-                        tracing::warn!(
-                            cred = %reclaimed.credential_id,
-                            ?outcome,
-                            "CredentialEvent::ReauthRequired publish dropped"
-                        );
-                    }
-                }
-            },
-        }
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{
-        sync::{
-            Arc,
-            atomic::{AtomicU32, Ordering},
-        },
-        time::Duration,
-    };
-
-    use nebula_core::CredentialId;
-    use nebula_eventbus::EventBus;
-    use nebula_storage::credential::{
-        ClaimAttempt, InMemoryRefreshClaimRepo, RefreshClaimRepo, ReplicaId,
-    };
-
-    use super::*;
-    use crate::credential::refresh::{
-        coordinator::{RefreshCoordConfig, RefreshCoordinator},
-        sentinel::{SentinelThresholdConfig, SentinelTrigger},
-    };
-
-    /// Build a [`RefreshCoordinator`] with a tight 30ms reclaim cadence
-    /// for spawn-path unit tests. Other invariants are scaled
-    /// proportionally so `validate()` accepts the shape.
-    fn fast_coord_for_test(repo: Arc<dyn RefreshClaimRepo>) -> Arc<RefreshCoordinator> {
-        Arc::new(
-            RefreshCoordinator::new_with(
-                repo,
-                ReplicaId::new("reclaim-spawn-test"),
-                RefreshCoordConfig {
-                    claim_ttl: Duration::from_millis(100),
-                    heartbeat_interval: Duration::from_millis(30),
-                    refresh_timeout: Duration::from_millis(30),
-                    reclaim_sweep_interval: Duration::from_millis(30),
-                    sentinel_threshold: 3,
-                    sentinel_window: Duration::from_hours(1),
-                },
-            )
-            .expect("validated fast-cadence config"),
-        )
-    }
-
-    /// Set up an expired RefreshInFlight claim so the next reclaim_stuck
-    /// returns it as a sentinel event.
-    async fn seed_stuck_inflight_claim(repo: &Arc<dyn RefreshClaimRepo>, cid: CredentialId) {
-        let claim = match repo
-            .try_claim(&cid, &ReplicaId::new("crashed"), Duration::from_millis(20))
-            .await
-            .expect("try_claim ok")
-        {
-            ClaimAttempt::Acquired(c) => c,
-            ClaimAttempt::Contended { .. } => panic!("setup must always acquire"),
-        };
-        repo.mark_sentinel(&claim.token)
-            .await
-            .expect("mark sentinel");
-        // Wait past the TTL.
-        tokio::time::sleep(Duration::from_millis(50)).await;
-    }
-
-    #[tokio::test]
-    async fn run_one_sweep_below_threshold_does_not_publish_reauth() {
-        let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-        let sentinel = Arc::new(SentinelTrigger::new(
-            Arc::clone(&repo),
-            SentinelThresholdConfig::default(),
-        ));
-        let bus = Arc::new(EventBus::<CredentialEvent>::new(16));
-        let mut subscriber = bus.subscribe();
-
-        // Seed one stuck RefreshInFlight claim. With the default
-        // threshold of 3, one event must NOT trigger ReauthRequired.
-        let cid = CredentialId::new();
-        seed_stuck_inflight_claim(&repo, cid).await;
-
-        let metrics = RefreshCoordMetrics::for_tests();
-        run_one_sweep(&repo, &sentinel, Some(&bus), &metrics, None)
-            .await
-            .expect("sweep ok");
-
-        // Subscriber should observe no events (we published none).
-        let recv = tokio::time::timeout(Duration::from_millis(50), subscriber.recv()).await;
-        assert!(
-            recv.is_err(),
-            "below-threshold sentinel must NOT publish ReauthRequired; got {recv:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_one_sweep_at_threshold_publishes_reauth_with_sentinel_reason() {
-        let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-        let sentinel = Arc::new(SentinelTrigger::new(
-            Arc::clone(&repo),
-            SentinelThresholdConfig::default(),
-        ));
-        let bus = Arc::new(EventBus::<CredentialEvent>::new(16));
-        let mut subscriber = bus.subscribe();
-
-        let cid = CredentialId::new();
-        let metrics = RefreshCoordMetrics::for_tests();
-        // Three sweeps, each seeded with a stuck RefreshInFlight claim
-        // for the same credential. The third call to
-        // `on_sentinel_detected` reaches threshold (default 3-in-1h)
-        // and the sweep must publish CredentialEvent::ReauthRequired.
-        for _ in 0..3 {
-            seed_stuck_inflight_claim(&repo, cid).await;
-            run_one_sweep(&repo, &sentinel, Some(&bus), &metrics, None)
-                .await
-                .expect("sweep ok");
-        }
-
-        // Drain the subscriber: there must be exactly one
-        // ReauthRequired event published on the third sweep.
-        let mut reauth_count = 0u32;
-        let mut last_reason: Option<ReauthReason> = None;
-        while let Ok(Some(event)) =
-            tokio::time::timeout(Duration::from_millis(50), subscriber.recv()).await
-        {
-            if let CredentialEvent::ReauthRequired {
-                credential_id,
-                reason,
-            } = event
-            {
-                assert_eq!(credential_id, cid);
-                reauth_count += 1;
-                last_reason = Some(reason);
-            }
-        }
-        assert_eq!(
-            reauth_count, 1,
-            "exactly one ReauthRequired event at threshold"
-        );
-        assert!(
-            matches!(
-                last_reason,
-                Some(ReauthReason::SentinelRepeated {
-                    event_count: 3,
-                    window_secs: 3600,
-                })
-            ),
-            "reason must be SentinelRepeated; got {last_reason:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_one_sweep_two_in_window_one_outside_does_not_escalate() {
-        // Custom config: threshold 2 with a tight 100ms window so we
-        // can stage "1 outside the window, 2 inside" deterministically.
-        let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-        let sentinel = Arc::new(SentinelTrigger::new(
-            Arc::clone(&repo),
-            SentinelThresholdConfig {
-                threshold: 2,
-                window: Duration::from_millis(100),
-            },
-        ));
-        let bus = Arc::new(EventBus::<CredentialEvent>::new(16));
-        let mut subscriber = bus.subscribe();
-
-        let cid = CredentialId::new();
-
-        let metrics = RefreshCoordMetrics::for_tests();
-
-        // Event #1.
-        seed_stuck_inflight_claim(&repo, cid).await;
-        run_one_sweep(&repo, &sentinel, Some(&bus), &metrics, None)
-            .await
-            .expect("sweep 1 ok");
-
-        // Wait past the rolling window so event #1 ages out.
-        tokio::time::sleep(Duration::from_millis(150)).await;
-
-        // Event #2 — first inside the new window.
-        seed_stuck_inflight_claim(&repo, cid).await;
-        run_one_sweep(&repo, &sentinel, Some(&bus), &metrics, None)
-            .await
-            .expect("sweep 2 ok");
-
-        // Drain the subscriber. With 1 event aged-out and 1 event
-        // inside (count=1, threshold=2), no ReauthRequired must have
-        // been published.
-        let mut reauth_count = 0u32;
-        while let Ok(Some(event)) =
-            tokio::time::timeout(Duration::from_millis(50), subscriber.recv()).await
-        {
-            if let CredentialEvent::ReauthRequired { .. } = event {
-                reauth_count += 1;
-            }
-        }
-        assert_eq!(
-            reauth_count, 0,
-            "events outside the rolling window must not push us over the threshold"
-        );
-    }
-
-    #[tokio::test]
-    async fn spawn_emits_reauth_after_threshold_at_cadence() {
-        // Black-box smoke: spawn the actual task with a 30ms cadence,
-        // seed three stuck rows over time, and observe the
-        // ReauthRequired emission on the bus.
-        let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-        let sentinel = Arc::new(SentinelTrigger::new(
-            Arc::clone(&repo),
-            SentinelThresholdConfig::default(),
-        ));
-        let bus = Arc::new(EventBus::<CredentialEvent>::new(16));
-        let mut subscriber = bus.subscribe();
-
-        let coord = fast_coord_for_test(Arc::clone(&repo));
-        let _handle = ReclaimSweepHandle::spawn(
-            Arc::clone(&coord),
-            Arc::clone(&sentinel),
-            Some(Arc::clone(&bus)),
-        );
-
-        let cid = CredentialId::new();
-        // Seed three stuck claims with brief gaps to let the sweep
-        // process each in turn.
-        let observed = Arc::new(AtomicU32::new(0));
-        let observed_clone = Arc::clone(&observed);
-        let cid_for_listener = cid;
-        let listener = tokio::spawn(async move {
-            while let Some(event) = subscriber.recv().await {
-                if matches!(
-                    event,
-                    CredentialEvent::ReauthRequired { credential_id, .. }
-                    if credential_id == cid_for_listener
-                ) {
-                    observed_clone.fetch_add(1, Ordering::SeqCst);
-                    break;
-                }
-            }
-        });
-
-        for _ in 0..3 {
-            seed_stuck_inflight_claim(&repo, cid).await;
-            // Give the sweep loop one cadence to pick this up before
-            // we seed the next.
-            tokio::time::sleep(Duration::from_millis(80)).await;
-        }
-
-        // Wait up to 1s for the listener to observe the event.
-        let _ = tokio::time::timeout(Duration::from_secs(1), listener).await;
-        assert_eq!(
-            observed.load(Ordering::SeqCst),
-            1,
-            "spawned reclaim sweep must publish exactly one ReauthRequired at threshold"
-        );
-    }
-
-    /// Sub-spec emission test for reclaim metrics.
-    ///
-    /// Calls `run_one_sweep` against a repo with one stuck
-    /// `RefreshInFlight` claim and asserts `reclaim_sweeps_total{
-    /// outcome=reclaimed}` and `sentinel_events_total{action=recorded}`
-    /// each tick once. A second sweep over an empty repo must tick
-    /// `reclaim_sweeps_total{outcome=no_work}`.
-    #[tokio::test]
-    async fn run_one_sweep_increments_metrics_per_sub_spec_6() {
-        use nebula_metrics::MetricsRegistry;
-
-        let registry = MetricsRegistry::new();
-        let metrics = RefreshCoordMetrics::with_registry(&registry).unwrap();
-        let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-        let sentinel = Arc::new(SentinelTrigger::new(
-            Arc::clone(&repo),
-            SentinelThresholdConfig::default(),
-        ));
-
-        let cid = CredentialId::new();
-        seed_stuck_inflight_claim(&repo, cid).await;
-
-        run_one_sweep(&repo, &sentinel, None, &metrics, None)
-            .await
-            .expect("first sweep ok");
-
-        assert_eq!(
-            metrics.reclaim_reclaimed.get(),
-            1,
-            "reclaim_reclaimed must tick once when stuck rows exist"
-        );
-        assert_eq!(
-            metrics.reclaim_no_work.get(),
-            0,
-            "reclaim_no_work must NOT tick when stuck rows exist"
-        );
-        assert_eq!(
-            metrics.sentinel_recorded.get(),
-            1,
-            "sentinel_recorded must tick once for the RefreshInFlight row"
-        );
-        assert_eq!(
-            metrics.sentinel_reauth_triggered.get(),
-            0,
-            "no escalation below threshold (first event of three)"
-        );
-
-        // Second sweep with no work must increment no_work, leaving
-        // the other counters untouched.
-        run_one_sweep(&repo, &sentinel, None, &metrics, None)
-            .await
-            .expect("second sweep ok");
-        assert_eq!(
-            metrics.reclaim_no_work.get(),
-            1,
-            "second sweep over empty repo must increment no_work"
-        );
-        assert_eq!(
-            metrics.reclaim_reclaimed.get(),
-            1,
-            "no further work; reclaimed counter must remain at 1"
-        );
-    }
-
-    /// Sub-spec — at-threshold sweep must increment both
-    /// `recorded` and `reauth_triggered`.
-    #[tokio::test]
-    async fn run_one_sweep_at_threshold_increments_reauth_triggered() {
-        use nebula_metrics::MetricsRegistry;
-
-        let registry = MetricsRegistry::new();
-        let metrics = RefreshCoordMetrics::with_registry(&registry).unwrap();
-        let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-        let sentinel = Arc::new(SentinelTrigger::new(
-            Arc::clone(&repo),
-            SentinelThresholdConfig::default(),
-        ));
-
-        let cid = CredentialId::new();
-        for _ in 0..3 {
-            seed_stuck_inflight_claim(&repo, cid).await;
-            run_one_sweep(&repo, &sentinel, None, &metrics, None)
-                .await
-                .expect("sweep ok");
-        }
-
-        assert_eq!(
-            metrics.sentinel_recorded.get(),
-            3,
-            "three sentinel events recorded"
-        );
-        assert_eq!(
-            metrics.sentinel_reauth_triggered.get(),
-            1,
-            "exactly one reauth_triggered tick when threshold crossed"
-        );
-    }
-}
