+++ crates/engine/tests/credential_thundering_herd_tests.rs
@@ -120,7 +120,7 @@ async fn only_one_refresh_under_concurrent_access() {
     };
     store.put(cred, PutMode::CreateOnly).await.unwrap();
 
-    let resolver = Arc::new(nebula_engine::credential::CredentialResolver::new(store));
+    let resolver = Arc::new(nebula_engine::credential::CredentialResolver::new(store).unwrap());
     let ctx = CredentialContext::for_test("test-user");
 
     let mut handles = Vec::with_capacity(10);
