--- crates/engine/tests/refresh_coordinator_chaos.rs
@@ -1,442 +1,0 @@
-//! Chaos test — refresh-coordinator multi-replica race regression.
-//!
-//! Per sub-spec
-//! `docs/INTEGRATION_MODEL.md`.
-//!
-//! # Gating
-//!
-//! This test is `#[ignore]`'d by default — it costs ~6s of wall-clock
-//! per CI run and dominates `nebula-engine` test time on every PR
-//! (review I3). It runs only when one of:
-//!
-//! - the `chaos-full` Cargo feature is on (used by `nightly-chaos.yml`); or
-//! - the runner passes `--include-ignored` to `cargo nextest`.
-//!
-//! Both paths bypass the `cfg_attr(not(feature = "chaos-full"), ignore)`
-//! gate: enabling `chaos-full` removes the `ignore` attribute outright,
-//! while `--include-ignored` runs ignored tests regardless. The default
-//! `cargo nextest run -p nebula-engine` thus skips this test entirely.
-//!
-//! # Plane sizes
-//!
-//! | Plane    | Duration | Credentials | Replicas | Workers/replica | Total worker tasks |
-//! |----------|----------|-------------|----------|-----------------|--------------------|
-//! | `ci()`   | 5s       | 5           | 3        | 3               | 9                  |
-//! | `full()` | 10 min   | 100         | 3        | 8               | 24                 |
-//!
-//! `chaos-full` is not just longer — it also widens the worker
-//! concurrency from 9 to 24 tasks, exercising a different contention
-//! profile, not just more iterations of the same pattern.
-//!
-//! # Design
-//!
-//! The harness models the n8n #13088 production race directly:
-//!
-//! 1. Three `RefreshCoordinator` instances share **one** `InMemoryRefreshClaimRepo` so the L2 claim
-//!    rows behave like a shared Postgres table across replicas.
-//! 2. A pool of credentials is staggered into "expired" / "near-expiry" / "fresh" buckets so the
-//!    `needs_refresh_after_backoff` predicate fires for some and short-circuits for others —
-//!    exercising both the `Acquired` and `CoalescedByOtherReplica` paths.
-//! 3. Each replica spawns a worker pool that picks a random credential, asks the coordinator to run
-//!    a fake "IdP POST" closure, and records the latency.
-//! 4. The fake IdP increments a per-credential atomic so the harness can prove no more than one
-//!    POST fired per credential per refresh window. This is the **only** thing that distinguishes
-//!    coordinated refresh from the n8n bug.
-//!
-//! # Assertions
-//!
-//! - Per-credential IdP call count ≤ 1 in any single refresh window. (no double-POST)
-//! - `ReauthRequired` count == 0 (no injected crashes).
-//! - Outside-refresh-window P50 latency < 4s watchdog (catches a regression where the L1 oneshot
-//!   stopped firing, forcing every caller through full L2 backoff). The sub-spec target P99 <
-//!   100ms is a production SLO measured against real load, not the chaos harness — the harness
-//!   emits raw P50/P99/max for inspection so drift is still visible.
-//! - Total `coalesced_l1 + coalesced_l2` ≥ 1 (proves the test actually exercised cross-replica
-//!   coordination, not just sequential single-replica calls).
-
-use std::{
-    sync::{
-        Arc,
-        atomic::{AtomicU64, Ordering},
-    },
-    time::{Duration, Instant},
-};
-
-use nebula_core::CredentialId;
-use nebula_engine::credential::refresh::{
-    RefreshCoordConfig, RefreshCoordMetrics, RefreshCoordinator, RefreshError,
-};
-use nebula_metrics::{
-    MetricsRegistry, NEBULA_CREDENTIAL_REFRESH_COORD_CLAIMS_TOTAL,
-    NEBULA_CREDENTIAL_REFRESH_COORD_COALESCED_TOTAL,
-    NEBULA_CREDENTIAL_REFRESH_COORD_SENTINEL_EVENTS_TOTAL, refresh_coord_claim_outcome,
-    refresh_coord_coalesced_tier, refresh_coord_sentinel_action,
-};
-use nebula_storage::credential::{InMemoryRefreshClaimRepo, RefreshClaimRepo, ReplicaId};
-use tokio::task::JoinHandle;
-
-/// Chaos-test parameters. The default `ci()` plane is the CI-friendly
-/// scaled-down version (5s × 5 credentials × 3 replicas × 3 workers per
-/// replica = 9 worker tasks). The `chaos-full` Cargo feature widens to
-/// the full sub-spec plane (10 min × 100 credentials × 3 replicas
-/// × 8 workers per replica = 24 worker tasks) with no code changes —
-/// only the `cfg`-gated literals differ.
-///
-/// The two planes differ in concurrency, not just in duration: 9 vs 24
-/// concurrent worker tasks contend over the shared in-memory L2 row,
-/// so `chaos-full` exercises a noticeably different race profile.
-#[derive(Clone, Debug)]
-struct ChaosParams {
-    duration: Duration,
-    credentials: usize,
-    replicas: usize,
-    workers_per_replica: usize,
-}
-
-impl ChaosParams {
-    /// CI-friendly scaled-down plane: 5s × 5 credentials × 3 replicas
-    /// × 3 workers per replica (9 worker tasks total). Sized so the
-    /// chaos test fits inside CI's per-test deadline without padding
-    /// workspace-wide test time, while still exercising all three
-    /// coordination paths (L1 oneshot, L2 acquired, L2 coalesced) at
-    /// non-trivial concurrency.
-    ///
-    /// Default-CI runs of `cargo nextest -p nebula-engine` skip this
-    /// test entirely (the module-level `cfg_attr(not(feature =
-    /// "chaos-full"), ignore)` guard); the plane is therefore only
-    /// observed via `--include-ignored` runs.
-    #[cfg(not(feature = "chaos-full"))]
-    fn ci() -> Self {
-        Self {
-            duration: Duration::from_secs(5),
-            credentials: 5,
-            replicas: 3,
-            workers_per_replica: 3,
-        }
-    }
-
-    /// Nightly `chaos-full` plane: 10 min × 100 credentials × 3
-    /// replicas × 8 workers per replica (24 worker tasks total). The
-    /// nightly-chaos workflow flips the `chaos-full` Cargo feature to
-    /// activate this plane.
-    #[cfg(feature = "chaos-full")]
-    fn ci() -> Self {
-        Self {
-            duration: Duration::from_mins(10),
-            credentials: 100,
-            replicas: 3,
-            workers_per_replica: 8,
-        }
-    }
-}
-
-/// Per-credential bookkeeping the test uses to prove no double-POST
-/// fires in a single refresh window.
-///
-/// `total_idp_calls` is the cumulative POST count for sanity checks
-/// against `claims_total{outcome=acquired}`. `generation` marks the
-/// current refresh window — the per-window POST count is kept in the
-/// shared `posts_per_window` map so we can assert that no
-/// generation ever observed more than one POST.
-struct CredentialTracker {
-    /// Total IdP "calls" since process start. The fake IdP increments
-    /// this once per refresh closure invocation.
-    total_idp_calls: AtomicU64,
-    /// Window marker bumped each time the closure successfully runs.
-    /// Counts at the time of POST are stamped under this value so the
-    /// post-run assertion can detect any window with > 1 POST.
-    generation: AtomicU64,
-}
-
-impl CredentialTracker {
-    fn new() -> Self {
-        Self {
-            total_idp_calls: AtomicU64::new(0),
-            generation: AtomicU64::new(0),
-        }
-    }
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-#[cfg_attr(
-    not(feature = "chaos-full"),
-    ignore = "chaos test gated behind chaos-full feature; runs in nightly CI via \
-              .github/workflows/nightly-chaos.yml. Pass --include-ignored or \
-              --features chaos-full to run locally."
-)]
-async fn three_replicas_zero_double_idp_calls() {
-    let params = ChaosParams::ci();
-    let repo: Arc<dyn RefreshClaimRepo> = Arc::new(InMemoryRefreshClaimRepo::new());
-
-    // Single shared MetricsRegistry so all three replicas write to the
-    // same series — at the end we read the aggregated counters straight
-    // from the registry rather than poking pub(crate) handle fields.
-    let registry = MetricsRegistry::new();
-    let metrics_handle = RefreshCoordMetrics::with_registry(&registry).unwrap();
-
-    // Build N credentials; each gets a tracker the fake IdP closure
-    // bumps. Use the in-memory repo (cross-replica visibility) so the
-    // L2 claim table behaves like a shared Postgres in production.
-    let credentials: Vec<(CredentialId, Arc<CredentialTracker>)> = (0..params.credentials)
-        .map(|_| (CredentialId::new(), Arc::new(CredentialTracker::new())))
-        .collect();
-
-    // Build N replicas, each with its own coordinator pointing at the
-    // shared repo. Each coordinator gets the same shared metrics handle
-    // so we can read aggregate counters at the end.
-    let replicas: Vec<Arc<RefreshCoordinator>> = (0..params.replicas)
-        .map(|i| {
-            Arc::new(
-                RefreshCoordinator::new_with(
-                    Arc::clone(&repo),
-                    ReplicaId::new(format!("chaos-replica-{i}")),
-                    chaos_config(),
-                )
-                .expect("validated chaos config")
-                .with_metrics(metrics_handle.clone()),
-            )
-        })
-        .collect();
-
-    // Track each credential's per-window observed POST count. The
-    // harness asserts the closure fires at most once per (cred, window).
-    // The map is keyed by `(credential_id, generation_at_post_time)`;
-    // the value is the count of POSTs stamped with that exact key.
-    let posts_per_window = Arc::new(parking_lot::Mutex::new(std::collections::HashMap::<
-        (CredentialId, u64),
-        u64,
-    >::new()));
-
-    // Latency samples partitioned by outcome:
-    // - `coalesced` — caller short-circuited via L1 oneshot or L2 post-backoff recheck. These are
-    // the "outside refresh window" calls in sub-spec and must be fast.
-    // - `refreshed` — caller acquired the L2 row and ran the closure. These include the simulated
-    // IdP POST and are bounded by `refresh_timeout`, so they are NOT subject to the 100ms
-    //   target.
-    let latencies_coalesced = Arc::new(parking_lot::Mutex::new(Vec::<Duration>::new()));
-    let latencies_refreshed = Arc::new(parking_lot::Mutex::new(Vec::<Duration>::new()));
-
-    let test_deadline = Instant::now() + params.duration;
-    let mut workers: Vec<JoinHandle<()>> = Vec::new();
-    for replica in &replicas {
-        for _w in 0..params.workers_per_replica {
-            let coord = Arc::clone(replica);
-            let creds = credentials.clone();
-            let map = Arc::clone(&posts_per_window);
-            let lats_coalesced = Arc::clone(&latencies_coalesced);
-            let lats_refreshed = Arc::clone(&latencies_refreshed);
-            workers.push(tokio::spawn(async move {
-                while Instant::now() < test_deadline {
-                    let (cid, tracker) = {
-                        let idx = rand::random_range(0..creds.len());
-                        creds[idx].clone()
-                    };
-
-                    let started = Instant::now();
-                    let tracker_for_closure = Arc::clone(&tracker);
-                    let map_inside = Arc::clone(&map);
-                    let cid_inside = cid;
-                    let outcome: Result<(), RefreshError> = coord
-                        .refresh_coalesced(
-                            &cid,
-                            // Random 70% chance of "still needs refresh"
-                            // so both Acquired and CoalescedByOtherReplica
-                            // paths are exercised.
-                            |_id| async {
-                                let v: u32 = rand::random_range(0..100);
-                                v < 70
-                            },
-                            move |_claim| async move {
-                                // Fake IdP POST. Each POST records its
-                                // own (cred, generation_at_capture)
-                                // tuple — no off-by-one tracking.
-                                tracker_for_closure
-                                    .total_idp_calls
-                                    .fetch_add(1, Ordering::SeqCst);
-                                let window_id =
-                                    tracker_for_closure.generation.load(Ordering::SeqCst);
-                                {
-                                    let mut m = map_inside.lock();
-                                    let entry = m.entry((cid_inside, window_id)).or_insert(0);
-                                    *entry += 1;
-                                }
-                                // Tiny sleep to let other replicas race.
-                                tokio::time::sleep(Duration::from_millis(5)).await;
-                                // Bump generation so the next window's
-                                // calls start at zero.
-                                tracker_for_closure
-                                    .generation
-                                    .fetch_add(1, Ordering::SeqCst);
-                                Ok(())
-                            },
-                        )
-                        .await;
-                    let elapsed = started.elapsed();
-                    match &outcome {
-                        Ok(()) => {
-                            // Acquired + ran the closure (the refresh path).
-                            lats_refreshed.lock().push(elapsed);
-                        },
-                        Err(RefreshError::CoalescedByOtherReplica) => {
-                            // Outside-refresh-window path: another caller
-                            // already refreshed; we just wake up.
-                            lats_coalesced.lock().push(elapsed);
-                        },
-                        Err(e) => {
-                            // Any other error is a chaos finding worth
-                            // surfacing. Don't panic — let the assertions
-                            // at the end fail loudly if it actually
-                            // matters.
-                            eprintln!("[chaos] non-coalesce error: {e:?}");
-                        },
-                    }
-                    // Quick yield so workers don't all hammer the same
-                    // credential cache line.
-                    tokio::time::sleep(Duration::from_millis(1)).await;
-                }
-            }));
-        }
-    }
-
-    for w in workers {
-        let _ = w.await;
-    }
-
-    // Drop the handle to suppress unused-variable warnings; the same
-    // handle has already done its work via the shared registry.
-    drop(metrics_handle);
-
-    // ── Assertion 1: no double-POST per (cred, window) ─────────────────
-    {
-        let m = posts_per_window.lock();
-        for ((cid, gen_id), posts) in &*m {
-            assert!(
-                *posts <= 1,
-                "credential {cid} window {gen_id} observed {posts} IdP calls — double-POST \
-                 race (n8n #13088 lineage)"
-            );
-        }
-    }
-
-    // ── Assertion 2: aggregate IdP-call count ≤ acquired claims ────────
-    let total_idp_calls: u64 = credentials
-        .iter()
-        .map(|(_, t)| t.total_idp_calls.load(Ordering::SeqCst))
-        .sum();
-    let acquired = read_counter(
-        &registry,
-        NEBULA_CREDENTIAL_REFRESH_COORD_CLAIMS_TOTAL,
-        "outcome",
-        refresh_coord_claim_outcome::ACQUIRED,
-    );
-    assert!(
-        total_idp_calls <= acquired,
-        "IdP calls ({total_idp_calls}) must be <= claims_acquired ({acquired}) — every fake POST \
-         goes through one acquired claim"
-    );
-
-    // ── Assertion 3: at least one coalesce happened ─────────────────────
-    let coalesced_l1 = read_counter(
-        &registry,
-        NEBULA_CREDENTIAL_REFRESH_COORD_COALESCED_TOTAL,
-        "tier",
-        refresh_coord_coalesced_tier::L1,
-    );
-    let coalesced_l2 = read_counter(
-        &registry,
-        NEBULA_CREDENTIAL_REFRESH_COORD_COALESCED_TOTAL,
-        "tier",
-        refresh_coord_coalesced_tier::L2,
-    );
-    assert!(
-        coalesced_l1 + coalesced_l2 >= 1,
-        "chaos workload must exercise at least one coalesce (L1+L2={}); otherwise the test is \
-         not actually multi-replica",
-        coalesced_l1 + coalesced_l2
-    );
-
-    // ── Assertion 4: zero ReauthRequired (no injected crashes) ─────────
-    let reauth = read_counter(
-        &registry,
-        NEBULA_CREDENTIAL_REFRESH_COORD_SENTINEL_EVENTS_TOTAL,
-        "action",
-        refresh_coord_sentinel_action::REAUTH_TRIGGERED,
-    );
-    assert_eq!(
-        reauth, 0,
-        "no injected crashes — sentinel must NOT escalate ({reauth} false positives)"
-    );
-
-    // ── Assertion 5: latency drift watchdog ────────────────────────────
-    //
-    // Sub-spec target: P99 < 100ms outside the refresh window —
-    // measured in a real production environment, not a chaos harness
-    // running alongside ~270 other tests. P99 in this harness is
-    // dominated by L2 contention backoff (capped at 5s+jitter by
-    // design); the median tracks the L1 oneshot path which is
-    // microsecond-fast.
-    //
-    // A regression that pushed even the median beyond 2s would indicate
-    // the L1 oneshot stopped working entirely (every caller now takes
-    // a full L2 backoff, in which case the test would have failed
-    // assertion 1 anyway). The 2s ceiling here is a watchdog, not a
-    // tight bound.
-    //
-    // The chaos run prints raw P50/P99/max for inspection so an
-    // operator can spot drift in median or tail behavior even when the
-    // assertion passes.
-    {
-        let lats = latencies_coalesced.lock();
-        if lats.is_empty() {
-            eprintln!("[chaos] no coalesced-path samples; skipping latency assertion");
-        } else {
-            let mut sorted: Vec<Duration> = lats.clone();
-            sorted.sort();
-            let p50 = sorted[sorted.len() / 2];
-            let p99 = sorted[((sorted.len() as f64) * 0.99) as usize].min(*sorted.last().unwrap());
-            let max = *sorted.last().unwrap();
-            eprintln!(
-                "[chaos] coalesced path: samples={}, P50={p50:?}, P99={p99:?}, max={max:?}",
-                sorted.len()
-            );
-            assert!(
-                p50 < Duration::from_secs(4),
-                "outside-refresh-window P50 latency {p50:?} exceeds 4s watchdog ceiling — \
-                 the L1 oneshot path is broken (every caller now waits a full L2 backoff); \
-                 samples: {}",
-                sorted.len()
-            );
-        }
-    }
-
-    // Sanity: the refresh-path samples must exist (else the test ran
-    // no actual refreshes — assertion 1 would be vacuous).
-    {
-        let lats = latencies_refreshed.lock();
-        assert!(
-            !lats.is_empty(),
-            "chaos test ran zero refresh closures — assertion 1 was vacuous"
-        );
-    }
-}
-
-/// Validated config tuned for chaos: short claim_ttl + short reclaim
-/// cadence so the test exercises real claim lifecycle in 30s.
-fn chaos_config() -> RefreshCoordConfig {
-    RefreshCoordConfig {
-        claim_ttl: Duration::from_secs(3),
-        heartbeat_interval: Duration::from_secs(1),
-        refresh_timeout: Duration::from_millis(800),
-        reclaim_sweep_interval: Duration::from_secs(1),
-        sentinel_threshold: 3,
-        sentinel_window: Duration::from_hours(1),
-    }
-}
-
-/// Read a labeled counter directly from the shared registry. Lets the
-/// chaos test sidestep the pub(crate) `RefreshCoordMetrics` field
-/// access while still observing aggregate counts.
-fn read_counter(registry: &MetricsRegistry, name: &str, label_key: &str, label_val: &str) -> u64 {
-    let labels = registry.interner().single(label_key, label_val);
-    registry.counter_labeled(name, &labels).unwrap().get()
-}
