+++ deny.toml
@@ -111,8 +111,9 @@ deny = [
     "nebula-tenancy",
     "nebula-engine",
     "nebula-api",
+    "nebula-credential",
     "nebula-credential-runtime",
-  ], reason = "Storage port is a Core-tier contract; the adapter, tenancy decorator, and exec/api consumers (incl. the credential-runtime facade, which derives the canonical Scope::credential_owner_id) may depend on it" },
+  ], reason = "Storage port is a Core-tier contract; the adapter, tenancy decorator, exec/api consumers, and nebula-credential (which hosts the refresh coordinator and depends on RefreshClaimStore, ADR-0092) may depend on it" },
 
   # nebula-tenancy is the Business-tier multi-tenancy security boundary: it
   # owns the policy (resolve `Scope` from a `Principal`, inject it, enforce
